pyproject.toml
[tool.poetry]
name = "brien"
version = "1"
description = "requirement"
authors = ["Your Name <you@example.com>"]
license = "n"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
amqp = "^5.1.1"
asgiref = "^3.7.2"
attrs = "^23.1.0"
billiard = "^4.1.0"
build = "^0.10.0"
cachetools = "^5.3.1"
certifi = "^2023.5.7"
chardet = "^5.1.0"
charset-normalizer = "^3.1.0"
cleo = "^2.0.1"
click = "^8.1.3"
click-didyoumean = "^0.3.0"
click-plugins = "^1.1.1"
click-repl = "^0.2.0"
colorama = "^0.4.6"
crashtest = "^0.4.1"
defusedxml = "^0.7.1"
diff-match-patch = "^20230430"
distlib = "^0.3.6"
django-environ = "^0.10.0"
django-extensions = "^3.2.1"
django-filter = "^23.2"
django-import-export = "^3.2.0"
django-rest-framework = "^0.1.0"
django-shell-plus = "^1.1.7"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
dulwich = "^0.21.5"
environ = "^1.0"
et-xmlfile = "^1.1.0"
filelock = "^3.12.0"
google-api-core = "^2.11.0"
google-auth = "^2.19.1"
google-crc32c = "^1.5.0"
googleapis-common-protos = "^1.59.0"
html5lib = "^1.1"
httplib2 = "^0.22.0"
idna = "^3.4"
importlib-metadata = "^6.6.0"
installer = "^0.7.0"
jaraco-classes = "^3.2.3"
jinja2 = "^3.1.2"
jsonschema = "^4.17.3"
keyring = "^23.13.1"
kombu = "^5.2.4"
lockfile = "^0.12.2"
markuppy = "^1.14"
markupsafe = "^2.1.3"
more-itertools = "^9.1.0"
msgpack = "^1.0.5"
oauthlib = "^3.2.2"
odfpy = "^1.4.1"
openpyxl = "^3.1.2"
packaging = "^23.1"
pexpect = "^4.8.0"
pkginfo = "^1.9.6"
platformdirs = "^3.5.1"
poetry-core = "^1.6.1"
prompt-toolkit = "^3.0.38"
protobuf = "^4.23.2"
ptyprocess = "^0.7.0"
pyasn1 = "^0.5.0"
pyasn1-modules = "^0.3.0"
pyjwt = "^2.7.0"
pyparsing = "^3.0.9"
pyproject-hooks = "^1.0.0"
pyrsistent = "^0.19.3"
pytz = "^2023.3"
pywin32-ctypes = "^0.2.0"
pyyaml = "^6.0"
redis = "^4.5.5"
requests = "^2.31.0"
requests-oauthlib = "^1.3.1"
requests-toolbelt = "^1.0.0"
rest-framework-simplejwt = "^0.0.2"
rsa = "^4.9"
shellingham = "^1.5.0.post1"
six = "^1.16.0"
sqlparse = "^0.4.4"
tablib = "^3.4.0"
tomlkit = "^0.11.8"
trove-classifiers = "^2023.5.24"
tzdata = "^2023.3"
vine = "^5.0.0"
virtualenv = "^20.23.0"
wcwidth = "^0.2.6"
webencodings = "^0.5.1"
xlrd = "^2.0.1"
xlwt = "^1.3.0"
zipp = "^3.15.0"
celery = "^5.2.7"
psycopg2 = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
